Change target to build locally
Changed the Makefile target to build locally. Am altering stf_ios_support to copy the binary from there instead of building into a relative directory which was confusing.

This commit also removes the shutdown check within the main video processing loop. It was causing problems so have commented it out for now.