Correct README
Corrected README. This repo is not in fact a modification of quicktime_video_hack in any way. It can be used together with frames sources that are a modification of that repo, but it is not itself a modification. ( hence license can be changed )

Also removed the link to MIT license since it is no longer under MIT license.